LayoutTests/imported/w3c/web-platform-tests/html/canvas/element/path-objects/2d.path.rect.winding-expected.txt

2d.path.rect.winding
Actual output:

FAIL Canvas test: 2d.path.rect.winding assert_equals: Red channel of the pixel at (25, 12) expected 0 but got 255
